ath12k: enable mlo capability by default
Currently, mlo capability of the driver is controlled via the mlo_capable
module param. The default value of this param is false which makes the
driver load with mlo capability disabled at boot time.

Change the default value to true in order to enable the mlo capability from
boot time itself.

Change-Id: I0e9e2f195a13e24dbf22be0b3a9356e1f0951ee3
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>