mac80211.sh: Script changes for channel switch
In repeater mode, to make AP interface switch to STA channel during
CSA, separate script is called. The script will be created in case
of 6GHz. When auto mode is configured, during APUT bringup, frequency
will not be known by the script, hence additional script to take care
of AP channel switch is not created which caused AP to not switch
channel.

To fix the issue, band is passed from mac80211.sh and used to create
the additional script for channel switch.
Hostapd cli will configure hw mode during AP bringup after channel
switch. hw mode is configured as "a" when channel number is greater than
36. When Root AP switches to 6Ghz channel less then 36 (say, 1, 5)
hw_mode is configured as "g" for 6GHz causing AP hostapd to fail.
Fixed the same by checking whether it is 6g band along with channel
check.

Change-Id: Ia993f8fcbf8f4b17fd9278d6d9fb902d5cc919c2
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>