ath12k: Fix link self peer delete
Currently, the mac address passed to FW for Peer delete
is the vif address which could be different for one of the
link vifs in case of MLO. Though FW might have not used the address for
validation currently, fix the address passed to the FW
for avoiding any future issues.

Also, many debug prints are based on vif->addr which is wrong,
this change fixes it as well.

Fixes: I94edc1185f857dbf ("ath12k: prepare vif and sta for MLO handling")
Change-Id: Ic4c58b34b32f320d7faddaf376e727e60964d916
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>