wifi: ath12k: fix 'link_stats' already exists error.
When firmware recovery is configured and simulate the firmware assert,
Once the firmware recovery is completed the following error is seen
on the console.

 debugfs: File 'link_stats' in directory 'netdev:wlan1' already present!

To handle this, add changes to create and delete the corresponding
debugfs file when interface add and remove is called respectively. Also,
do not create debugfs file if already present

Change-Id: I8a549b3dc6f5dfb23e05e811f42cb15ae40bbfff
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>