wifi: ath12k: fix usage of same vdev_id for same soc
when stability test is ran with multiple combination of
mesh interfaces, monitor interfaces and virtual AP interfaces,
firmware assert is seen with the reason of same vdev_id used
for two interfaces in same soc.

This is because, whenever vdev_id is assigned it is protected
by ar locks and the updation of free vdev is done only after wmi
command successful which are not sufficient in case of split mac cases,
where more than one radio is present in an soc, and it does not protect
under race condition where two interface simultaneously fetch the same
vdev id when one of the interface is still processing its wmi command.

To handle the same, add changes to protect the free_vdev_map with
ab locks as well as update the free_vdev_map as soon as fetched for
assignment without waiting for wmi command result.

Change-Id: Ia24dc8ae45f2d019262a4025cc58ed343829cd3c
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>