mac80211.sh: Restrict WDS config when not supported
Currently we are not checking if device is in native wifi
or in eth encap mode when "option wds 1" is added in config.

WDS AP should not bring up if device is in eth encap mode.

Currently WDS is only supported on native wifi mode,
hence restricting it on eth encap mode.

Change-Id: Ib4040a990f06cfa188f8f2ce9ebc697c834773a2
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>