ath12k: Fix filter setup for ext rx stats enable
Currently when ext_rx_stats is enabled, the stats are
not updated properly until the AP is restarted.
This is due to wrong configuration of the mac id during
filter setup when stats is enabled. It works later because
the filter is applied properly during driver start when the
interface is restarted.

Correct this by passing proper mac id to the filter setup function.

Fixes: I3a504281be47d("mac80211: Package Upgrade 08/18/2022")
Change-Id: Ieb22150005a4ba19c8eaea4bd05021f1786ef6dd
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>