ath12k: fix hw chan advertisement logic
Currently, regulatory and hardware capabilities were used to decide
the min and max frequency supported by a pdev. Based on this, no. of
channels available was being calculated from possible list of channels
in the particular band. This is wrong since if regulatory changes, then
the frequency limit can change for a pdev and that will not be taken
care owing to the fact that frequency update for a pdev is done only
once during mac registration. Also, taking the limit from regulatory or
hardware capability restricts the few channels from getting advertised
in phy info command.

Typically, the info command should list all channels available in the band
and then based on the regulatory and hardware capability, it should mark
those channels as enabled or disabled.

Fix this issue by using the band specific limits in pdev during registration.
Regulatory updates will take care of marking channels and enabled and disabled.

Fixes: 7ca3d5d4a845 ("ath12k: RDP454 split phy bringup changes")
Change-Id: Id55eeffab16038acf23737c2735d6250797c23c5
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>