ath12k: fix Tx power display during active CAC
Tx power is fetched from firmware's pdev stats. However, during active
CAC, the Tx power is not updated in the stats and is just filled with
the max allowed power. This leads to incorrect Tx power display on
the user space side.

Fix this issue by returning Tx power as 0 during active CAC since it
is known that dring CAC, there will be no transmission hapenning.

Fixes: 703d98bb171c ("ath12k: add get_txpower mac ops")
Change-Id: I72d768e9c0ca774a372420c3ed1a85690ebfb8b3
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>