ath12k: fix fw consumed wmi event for mgmt rx reordering
Currently firmware consumed wmi events are not handled, and so
there are chances of miss updating host snapshots.

Firmware consumed events are used to update the mgmt rx re-order
process, this info is essential in updating pkt counters properly, so
that mgmt rx reorder process will not go out-of-sync.

Handle firmware consumed WMI event and update the re-order process with
NULL skb.

Fixes: I98a7a2dd7c ("ath12k: add support for Mgmt Rx re-ordering")

Change-Id: If1039a18a0b0f128f91ce0f126bc68bfd64e9c0c
Signed-off-by: Karthikeyan Kathirvel <quic_kathirve@quicinc.com>