mac80211.sh: fix mesh + ML VAP bring up
Currently, when 2G+5G ML VAP  and 6G is configured as mesh,
then ML VAP is not coming up.

In case of ML VAP, the design is to bring up a single
instance of hostapd. Hence, we use the radio count which
are enabled. This is causing issue in case of the above
combination where the radio count is calculated as 3 when
one of the band does not have ap interface
Therefore, to start single hostapd instance we are waiting
for 3 event but actually we would receive only 2 event as only
they have the ap mode.

To handle this, change the logic to calculate the radio count
based on the ap mode interface.

Change-Id: I51f9e066102e2c6127fded22cb3605c4791e51bb
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>