ath12k: Fix send link mac address to get correct stats
Currently when we request for htt_peer_stats when SLO sta is
connected, we send mld mac address to get htt_peer_stats and
we are not getting any stats. In case of peer pktlog stats,
we are requesting for mld mac address and it should be
corrected.

To resolve this issue, we should send link sta mac address to get
proper stats.

Fixes: 193b858c62ab ("ath12k: move link specific dp stats to link sta directory")

Change-Id: I3ae58ebbdaa09c5e3b4f0f843b5b57c0d90039c7
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>