cfg80211/ath12k: fix wrong dfs cac time update not based on chandef
Currently, dfs_cac_time for agile DFS is calculated based on the channel
instead of chan_def. In case of channel 100, 160 MHz in ETSI domain,
Weather radar channels are not considered which results in wrong cac time.

Added the cfg80211 api to fetch the dfs_cac_time based
on the whole chandef.

Change-Id: I8f02fbf2c971e15fe28dc7ad50352439ff38cc49
Signed-off-by: Karthik M <quic_karm@quicinc.com>