ath12k: Fix inactive pdev access during mgmt Rx
Currently, though a mgmt Rx event is received from firmware
with a wrong pdev id in the event for which a corresponding
pdev in the ab's pdev list(in case of split phy) but is currently
stopped and inactive, we pass the corresponding mgmt frame to mac80211
which results in the below call trace since local->started flag
is false.

Call trace:

 ieee80211_rx_list+0xe4/0xa4c [mac80211]
 ieee80211_rx_napi+0x58/0xcc [mac80211]
 ath12k_hal_reo_shared_qaddr_cache_clear+0x1044/0x43b0 [ath12k]
 ath12k_wmi_free_tpc_stats_mem+0xae4/0x2948 [ath12k]
 ath12k_htc_rx_completion_handler+0x428/0x46c [ath12k]
 ath12k_ce_per_engine_service+0x264/0x324 [ath12k]
 ath12k_pci_hif_resume+0x20c/0x384 [ath12k]
 tasklet_action_common.isra.2+0xa4/0x11c
 tasklet_action+0x24/0x2c

Hence fetch ar during mgmt Rx only if the pdev is active.

Fixes: I5d3411533fe("ath12k: 11be spectral scan support")

Change-Id: Iaaeb3d254f20cf9e0db852750d7110daf4ef56e4
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>