ath12k: fix fw consumed pkt handling when mgmt rx reorder is disabled
In firmware consumed frame handling path of mgmt re-ordering, there is
no check for mgmt_rx_reorder and so when the feature is disabled this
path is getting executed unwantedly.

Below error messages are seen due to this issue

[  387.675744] ath12k_pci 0004:01:00.0: Failed to execute MGMT REO
reorder process
[  387.682711] ath12k_pci 0004:01:00.0: MGMT Rx REO timer is not
initialised
[  387.689799] ath12k_pci 0004:01:00.0: Failed to execute MGMT REO
reorder process
[  387.696754] ath12k_pci 0004:01:00.0: MGMT Rx REO timer is not
initialised
[  387.703870] ath12k_pci 0004:01:00.0: MGMT Rx REO timer is not
initialised

Add a feature check to deny executing the firmware consumed wmi events when
feature flag is disabled.

Fixes: If1039a18a0b ("ath12k: fix firmware consumed wmi event for mgmt rx reordering")

Change-Id: I0872a4bc0823a6c69b0b7d1e93d3b91ff3b18c6d
Signed-off-by: Karthikeyan Kathirvel <quic_kathirve@quicinc.com>