ath12k: Add support to identify dual mac
While debugging using cnss diag for dual mac device, we are using Data.msc
file to get the logs. Instead we should be using Data_dual.msc file for dual
mac devices to get the logs.

Add a check if the device is dual mac and send it through nl command to be
sure that correct .msc file is used.

Change-Id: Iaf729ea29ab216a17620568de8330d0e8ebfd2b5
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>