ath12k: Fix get peer from peer id instead of ast index
Currently in function ath12k_dp_mon_rx_update_user_stats, we are fetching peer
by ast index look up. But in monitor mode the ast index is maped to 0 to 15 bits
and in latest HTT_T2H_MSG_TYPE_PEER_MAP3 ast hash is mapped to 16 to 19 bits.
Due to this when multiple clients are connected we are not able to fetch peer
from ast index look up.

Peer ID is better way to access peer istead of ast index.

To fix this issue get peer from peer id.

Fixes: cd2d4e2d270a ("ath12k: Fix search flag and update peer map event for sta")

Change-Id: I28e064459632e91ccfc01a3e6c77e2caa858d5db
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>