ath12k: fix incorrect beacon template update during CSA
Once CSA finish is called, mac80211 changes the channel context via
reserved switch first and then schedules the CSA finish work which
then updates the link's CSA active state as well as the beacon. While
the reserved switch is hapenning, mac80211 calls the driver callback.
In this callback, ath12k performs the vdev restart and since it was
already up, it tries to install the beacon template again. However
this is wrong since at this point the beacon is not yet updated and
hence the beacon update will be the older beacon only.

Fix this issue by not installing the beacon template if CSA active
state is present in the link. When the beacon template is actually
updated, bss link info changed notify will take care to update the
beacon and also bring the vdev up.

Change-Id: Iaf4ebf2f291a1bccc0e5166b0fdfe68813d7035e
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>