Fix the incorrect value of station dump fields in SLO/MLO  mode
While connecting the AP with a station in SLO/MLO mode observe
the Dtim period and Beacon Interval as 0 in the station dump.

During the start AP, these fields are updated in the specific
link config. But when setting the station info, it's referring
to the sdata bss config. Due to this, observing the issue.

Fix it by updating these fields with corresponding link config data
per link.

Change-Id: Ic1c5e56f164b2bc99c6c370ef3aabbbdc7e9a448
Signed-off-by: Sivashankari Madhavan <quic_sivamadh@quicinc.com>