ath12k: Fix crashing of partner SoC when FW assert fails
Whenever a FW assert is received in case of MODE0, partner chip
assert is also provided by FW and we wait to force crash the system
until all the FW asserts are received. But in some cases if FW assert
is not received we have already added a fool proof mechanism to crash
partner chipsets. But this was not triggered due to wrong check of the
ATH12K_FLAG_RECOVERY flag. This flag is currently checked in the current
SoC on which the FW assert was received.

Rather, we need to check on partner SoC before sending a hang command.

Fixes: I87e30ca136b("ath12k: add support for MLO SSR Mode0")
Change-Id: I22eac08bef413fe05f0ba8be029af2bdaa66d52c
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>