mac80211.sh: Add frequency list for sta mode
Currently, in single wiphy architecture, when sta interface is brought up
in one particular band from OpenWrt config file, then it scans for
all the bands instead of one.

Because of this, when same ssid is broadcasted on 2 GHz Band and 5GHz band,
then sta selects the 2 GHz band.

In order to overcome the above issue, add the frequency list of the band to
the wpa_supplicant config file. Hence, it would scan only the configured band.

Change-Id: I3394678fa2f3dd8af32cef15f07fb2c877227b9a
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>