ath12k: fix max nss calculation for eht
when a mesh peer tries to associate, then we are seeing following
firmware assert

wlan_peer.c:2748 Assertion peer->peer_nss > 0 failed param0 :zero,param1 :zero,param2 :zero

Recently,for miami the tx and rx nss for 12-13 mcs is
updated as 2 and 0 respectively. Because of this,
while parsing the eht mcs and nss capability, the nss
for peer is derived as 0. Hence, in peer assoc command
0 is sent to FW and assert is seen.

To address this, add changes to fetch the maximum nss of the
supported mcs rates.

Change-Id: I3d28b2b10322e1b60bb5ebe0c3427e1348535d8b
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>