ath12k: Synchronize setting and testing flush flags
Usually when work threads carry out the process of
reconfiguring the chip which was asserted, a lot of other
things can happen from upper layers, requesting some
actions especially when the chip is down. Adding a check
for chip recovery might help at times, but it need not
always work because of this worker thread async.

This change will add synchronization will checking and
updating chip recovery flag.

Change-Id: I5d96268908c37f5a72c4308a9036ce08658d1135
Signed-off-by: Ramanathan Choodamani <quic_rchoodam@quicinc.com>
Signed-off-by: Sidhanta Sahu <quic_sidhanta@quicinc.com>