ath1Xk: add support for Tx Power insertion
For certain action frames like the TPC Report IE in the spectrum management
TPC Report action frame, and in the Radio Measurement Link Measurement
Report action frame there is a requirement to fill in the current
and Max Tx power of the device in the packet.

Add support to fill in these in the above packets. Driver advertises this
capability support via NL80211_FEATURE_TX_POWER_INSERTION feature flag.

For software encrypted case, PMF packets will be encrypted already hence
driver can't add value in this particular case.

While at it, revert the logic to advertise the Max Tx power to upper layer
since that can't be used further with singly wiphy architecture.

Change-Id: Ia517648bd2944e19fa2b8cc9e167a0d8ec28060b
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>