ath12k: Fix NULL ptr access during mutex unlock
When an arvif needs to be created on a new radio, we
delete the arvif in old radio and create on a new one.
But in ath12k_mac_vdev_delete() we set arvif->ar as NULL,
but this is accessed to unlock mutex and for dbg logging
though it is set to NULL in the vdev delete function.

Fix by taking a backup of the arvif->ar before calling the
ath12k_mac_vdev_delete:

Fixes: Ic7dcc913036("ath12k: Protect access of unlinked ath12k link vifs")
Fixes: I7fe6c123192("ath12k: Add single wiphy registration support")

Change-Id: I9b02fed61f7a69e475df91b71b19898d9e5234b9
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>