mac80211.sh/apsta_mode.sh: add support for ml link configuration
Currently, when fronthaul interface is configured as Multi-link
ap interface, ap-sta script is not able to configure the channel
corresponding to backhaul sta into the fronthaul ML AP.

This is because, the multi-link fronthaul ap interface has to be
configured with link id detail in the hostapd_cli command.

Add changes to fetch the link information of the multi-link ap
interface and then match the link config frequency details with the
sta interface compatible frequency list. If the match is found, then
the corresponding link interface is configured.

Currently, ap-sta script exits if it doesnot find the control interface
of ap. For multi-link fronthaul ap, the control interfaces are appended
with link. Hence, add changes to check multi-link ap control interface

Change-Id: Ibcf6913fc3a13369d341db2ea8a741f360d8017c
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>