ath11k: Fix correcting the series of the patch
Currently in the NUC(x86 + Pine) build process, all private patches
starting from 500 and above gets removed.

Due to this, this ath11k patch also get removed, and when the monitor
interface is being created and tried to bring up the monitor interface,
the system goes into the hung state.

So, renaming this private patch to correct series.

Fixes: c7a96bed16 ("UPSTREAM: wifi: ath11k: fix monitor mode bringup crash")

Change-Id: Ic061d5df9f15d5a15183a9af767d6bb0a5b24aa2
Signed-off-by: Rahul Bhattacharjee <quic_rbhattac@quicinc.com>