ath12k: use link conf for color collision detection check during Rx
Currently, during color collision detection check in Rx path, bss conf
is fetched using vif->bss_conf which is incorrect in case of MLO.
Further this bss_conf is used later to get the link id which can be
wrong and lead to WARN_ON().

Instead, use link_sdata to fetch the link bss_conf which is also present
there.

Fixes: 47480246689b ("cfg80211/mac80211: add MLO support for HE-BSS color change events")
Change-Id: I89a482a19c3550ac02b7fe63adbae3c06cd42fc7
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>