ath12k: Avoid disabling eht rate on a ML bss
When bitratemask command is set with EHT MCS disabled for all NSS
on a ML BSS, the peer assoc command is set to the firmware
with a non 11BE Phymode. This results is some assumptions
in FW going wrong leading to issues. To avoid them, do not
allow disabling EHT rates on a ML BSS.

Change-Id: If70469bd70376052c258c0b1369ea3580d9fd0dc
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>