ath12k: fix firmware assert during peer channel width change
Currently, during peer assoc, peer's current bandwidth is sent in the
peer flags. Due to this, during bandwidth upgrade/downgrade, there was
a possibility that the new channel width sent could be incompatible
with the peer_flags set earlier and hence firmware assert was seen.
In order to avoid this situation, change was added to send peer phymode
as well along with the channel width. However, in multi client situations,
this could lead to firmware's buffer getting fully exhausted which leads
to firmware assert.

Hence, send the max bandwidth in peer flags during peer assoc. This will
help in sending peer channel width alone whenver it changes and there is
no longer need of sending peer phmode with it. So, while at it, do code
cleanup related to sending peer phymode.

Change-Id: I5bd2ceb4e5c9140f42a6346eaaae3e09babd1494
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>