ath12k: skip WMI commands during ssr until wmi_init
If the FW is in the recovery process, we have to wait
till the first WMI command is sent, viz. WMI_INIT_CMDID.

At times firmware exception occurs with commands being sent before
WMI_INIT, when the FLUSH bit is cleared.

Avoid this by using a seperate flag for WMI, since
FLUSH flag is more of a generic flag.

Change-Id: Id0f5f0a1cd5df4cfd93109063a42f46b9ca7fe34
Signed-off-by: Ramanathan Choodamani <quic_rchoodam@quicinc.com>