ath12k: Force set set_fw_recovery of all abs in case of mlo capabale
It is expected to set ab->set_fw_recovery to true for all abs prior to firmware
recovery process. In case of mlo_capable >= 1, all partner chips per ag group
needs to be recovered irespective of the set_fw_recovery.

So, forcing the set_fw_recovery to true per ag when one of the
ab->set_fw_recovery is set to true.

Change-Id: Ia58ab3fd0d35d27c2c023b6a6d45f9df9d1ac5f5
Signed-off-by: Sidhanta Sahu <quic_sidhanta@quicinc.com>