ath12k: Add a failsafe mechanism  when FW recovery fails
Problem Statement:
When SSR fails to reconfigure the chip due to a host
or a FW issue, then the chip will not be in UP state.
It is neither recoverable nor re-assertable.

Solution:
In case SSR fails, it is better to issue a BUG_ON for
Mode0. This will crash the system and bring it to usable
state.

This change adds a module param to enable fail safe mechanism
when an external event disrupts the recovery process due to some
reason. The goal here is to keep the system in a usable state
due to any recovery disruption/failure..

Change-Id: I34940dd2475c4d39306493cf30f757e5b2061c54
Signed-off-by: Ramanathan Choodamani <quic_rchoodam@quicinc.com>
Signed-off-by: Sidhanta Sahu <quic_sidhanta@quicinc.com>