cfg80211/mac80211: fix scan issue due to single wiphy architecture
Currently, irrespective of the links active in an interface, if frequency
info is not provided in the scan request, all channels across all the
available bands under the wiphy are selected. This leads to huge scan request
as well as long time to complete the scan.

However, if an interface is active on a certain num of links, then there
is no point in scanning for all the bands and their channels.

Add fix to select channels only from bands which are currently active on
the requested scanning interface in such request.

While at it, also add check in driver to validate if vdev id bitmap is already
zero before calling ffs() on it.

Change-Id: I914c5a060e4b18d4316c608221392f7463b91bdb
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>