wifi: ath12k: Add SAWF module param check and modify logging of a log in SFE
Add SAWF module param check to avoid SAWF reconfiguration failure
logging, during firware recovery scenario as it is not expected when
SAWF feature is not enabled in driver.

Move the log print from ath12k_err to ath12k_dbg, in SFE as it is
needed only for debugging purpose.

Change-Id: I1cb52684c8a31424ce685cd0be782e3e9ed9551d
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>