ath12k: Disable qdss_tracing till recovery is complete
Problem statement:
When recovery is in progress, we use qdss_tracing flag
during coredump to populate QDSS_DATA region, but without
checking if we actually received a QMI event for QDSS_TRACE_MEM_REQ
before that. This will cause wrong offsets to be populated in the
seg_info if we never received that event mentioned above.

This scenario is likely to been hit when chip assertion happens when
reconfig is complete and we are yet to receive the
QMI_QDSS_TRACE_MEM_REQ event, thereby causing wrong seg_info offsets to be
populated during the coredump inside the reset thread. This will cause the
crash with below signature:
pc : __memcpy_fromio+0x30/0x78
lr : ath12k_spectral_get_dbring+0x1c4/0x1e4 [ath12k]

Solution:
As soon as the coredump occurs, check if the qdss_tracing was enabled
before and then disable it till recovery is complete. This flag is
set later when the QMI event received. So in case of a failure scenario
the disabled flag will continue to remain disabled.

Change-Id: I5f682244e7487057897c1d93ca06dbbab6fb477e
Signed-off-by: Ramanathan Choodamani <quic_rchoodam@quicinc.com>