ath11k : flush management frames to firmware before waiting for tx completion
warning print "ath11k c000000.wifi: failed to flush mgmt transmit queue 0"
is observed during interface down.

The management frames are queued in a skb_queue and the skb_queue
is dequeued in the work ar->wmi_mgmt_tx_work.

Before waiting for the tx completion of all the management frames,
we are not ensuring that the queued management frames are flushed to the
firmware.

This causes ar->num_pending_mgmt_tx to be positive and it leads to the
warning print.

Fix this by flushing all the management frames to firmware before waiting
for the tx completion.

Change-Id: I387e2184bb3959ded7343d29a1b32e5fa260b905
Signed-off-by: Hari Chandrakanthan <quic_haric@quicinc.com>