cfg80211: Fix bss list duplication and tree corruption
cfg80211 maintains scanned bsses in a list and in rbtree
for its usage. Its expected to have the bsses in both these
datastructures in sync. Having duplicate entries in the list
can result in tree corruption when the duplicate entry is tried
to be removed which is already not present in the tree.

Such a scenario happens during channel switch when the currently
associated bss node is updated in the rbtree. This update is done
by finding existing matching bss node from the tree with updated
channel after switch where it is unlinked from tree/list and the
current bss is readded after erasing. But if the erase is successful
but not the readd (which can happen due to cmp_bss() match if similar
entry is already available). This leads to inconsistency between tree
and list where the list has the entry but not the tree.
When this list is parsed in multiple places, say during scan dump or
bss expire, it comes across a rb node which is already deleted from
tree and rebalance/erase causes corruption.

The cause of bss_cmp() returning matching bss is due to current
implementation which returnes difference of channel frequency if
the channel pointers doesnt match. But in case of 6GHz, the channel
pointers can be different for different power modes but with same
frequency. This channel pointer update for the current bss happens
during channel switch which leads to the differences in ptr but when
frequencies match due to BW change the issue can happen.

To avoid this issue basically, instead of checking matching channel
pointers for 6GHz use the existing api to check if the channel is
same. This is done in this change.

Note:
As a followup, the rootcause of the issue which is setting a default
channel pointer during initial bss discovery should be done based
on pwr mode from the HE Oper IE from the discovered beacon so that
the pointer difference is not seen when chan switch happens.

Fixes: Ie109bf82e61("mac80211: Fix Channel Switch Announcement 6 GHz")

Change-Id: I13e79bc957fc581b56955c4baebddc7d0e901bf5
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>