wifi: ath12k: Add UMAC DP reset support
The code changes are to add support for umac dp
reset of partner SOC when one SOC crashes in MLO
as a part of MODE 1 SSR

In this process shared memory will be used by Host and
Firmware to communicate the sequence of events. Host will
clear the data structures such as tx/rx descriptors and
re-initialize the rings and reset hp/tp to 0.

Change-Id: I188c847ee3602feab6b5f682161af92beea2e89b
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>